Add EditextPreference to edit size for shape

+++ app/src/main/java/com/vnspectre/visualizerpreferences/AudioVisuals/VisualizerView.java
@@ -243,7 +243,7 @@ public class VisualizerView extends View {
      * Sets the scale for the minimum size of the shape.
      * @param scale the scale for the size of the shape.
      */
-    public void setMinSizeScale(int scale) {
+    public void setMinSizeScale(float scale) {
         TrailedShape.setMinSize(MIN_SIZE_DEFAULT * scale);
     }
